+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log of mruby uriparser
 
+- Added `uri.path` where `uri` is an instance of `URIParser::URI`.
+
 ## 0.1.0 - 2025-11-09
 
 - Added: `uri.host?`, `uri.dup`, `uri.clone`, `uri == other`, and per-component setter instance methods.